Split tests into two files, one for hashstore and one for hashfs_ext
