+++ tests/test_hashstore.py
@@ -103,7 +103,6 @@ def test_store_input_stream(pids, store):
         input_stream.close()
         ab_id = store.objects._get_sha256_hex_digest(pid)
         assert store.objects.exists(ab_id)
-    return
 
 
 def test_store_object_algorithm_args_invalid(store):
@@ -381,176 +380,3 @@ def test_get_sha256_hex_digest(pids, store):
     for pid in pids:
         hash_val = store.objects._get_sha256_hex_digest(pid)
         assert hash_val == pids[pid]["ab_id"]
-
-
-def test_computehash(pids, store):
-    test_dir = "tests/testdata/"
-    for pid in pids.keys():
-        path = test_dir + pid.replace("/", "_")
-        obj_stream = io.open(path, "rb")
-        obj_sha256_hash = store.objects.computehash(obj_stream, "sha256")
-        obj_stream.close()
-        assert pids[pid]["sha256"] == obj_sha256_hash
-
-
-def test_put(pids, store):
-    test_dir = "tests/testdata/"
-    for pid in pids.keys():
-        path = test_dir + pid.replace("/", "_")
-        hashaddress = store.objects.put(pid, path)
-        hashaddress_id = hashaddress.id
-        hashaddress_relpath = hashaddress.relpath
-        hashaddress_abspath = hashaddress.abspath
-        hashaddress_is_duplicate = hashaddress.is_duplicate
-        hashaddress_hex_digests = hashaddress.hex_digests
-        assert hashaddress_id == pids[pid]["ab_id"]
-        shard_id_path = "/".join(store.objects.shard(hashaddress_id))
-        assert hashaddress_relpath == shard_id_path
-        id_abs_path = store.objects.realpath(hashaddress_id)
-        assert hashaddress_abspath == id_abs_path
-        assert hashaddress_is_duplicate is False
-        assert hashaddress_hex_digests.get("md5") == pids[pid]["md5"]
-        assert hashaddress_hex_digests.get("sha1") == pids[pid]["sha1"]
-        assert hashaddress_hex_digests.get("sha256") == pids[pid]["sha256"]
-        assert hashaddress_hex_digests.get("sha384") == pids[pid]["sha384"]
-        assert hashaddress_hex_digests.get("sha512") == pids[pid]["sha512"]
-
-
-def test_put_additional_algorithm(pids, store):
-    test_dir = "tests/testdata/"
-    for pid in pids.keys():
-        algo = "sha224"
-        path = test_dir + pid.replace("/", "_")
-        hash_address = store.objects.put(pid, path, additional_algorithm=algo)
-        hex_digests = hash_address.hex_digests
-        sha224_hash = hex_digests.get(algo)
-        assert sha224_hash == pids[pid][algo]
-
-
-def test_put_with_correct_checksums(pids, store):
-    test_dir = "tests/testdata/"
-    for pid in pids.keys():
-        algo = "sha224"
-        algo_checksum = pids[pid][algo]
-        path = test_dir + pid.replace("/", "_")
-        store.objects.put(pid, path, checksum=algo_checksum, checksum_algorithm=algo)
-    assert store.objects.count() == 3
-
-
-def test_put_with_incorrect_checksum(pids, store):
-    test_dir = "tests/testdata/"
-    for pid in pids.keys():
-        algo = "sha224"
-        algo_checksum = "badChecksumValue"
-        path = test_dir + pid.replace("/", "_")
-        with pytest.raises(ValueError):
-            store.objects.put(pid, path, checksum=algo_checksum, checksum_algorithm=algo)
-    assert store.objects.count() == 0
-
-
-def test_move_and_get_checksums(pids, store):
-    test_dir = "tests/testdata/"
-    for pid in pids.keys():
-        path = test_dir + pid.replace("/", "_")
-        input_stream = io.open(path, "rb")
-        (
-            id,
-            hex_digests,
-            file_path,
-            is_duplicate,
-        ) = store.objects._move_and_get_checksums(pid, input_stream)
-        input_stream.close()
-        ab_id = store.objects._get_sha256_hex_digest(pid)
-        assert id == ab_id
-        assert hex_digests.get("md5") == pids[pid]["md5"]
-        assert hex_digests.get("sha1") == pids[pid]["sha1"]
-        assert hex_digests.get("sha256") == pids[pid]["sha256"]
-        assert hex_digests.get("sha384") == pids[pid]["sha384"]
-        assert hex_digests.get("sha512") == pids[pid]["sha512"]
-        assert os.path.isfile(file_path) is True
-        assert is_duplicate is False
-
-
-def test_move_and_get_checksums_duplicates(pids, store):
-    test_dir = "tests/testdata/"
-    for pid in pids.keys():
-        path = test_dir + pid.replace("/", "_")
-        input_stream = io.open(path, "rb")
-        store.objects._move_and_get_checksums(pid, input_stream)
-        input_stream.close()
-    for pid in pids.keys():
-        path = test_dir + pid.replace("/", "_")
-        input_stream = io.open(path, "rb")
-        (
-            id,
-            hex_digests,
-            file_path,
-            is_duplicate,
-        ) = store.objects._move_and_get_checksums(pid, input_stream)
-        input_stream.close()
-        assert is_duplicate is True
-        assert store.objects.count() == 3
-
-
-def test_mktempfile(pids, store):
-    test_dir = "tests/testdata/"
-    for pid in pids.keys():
-        path = test_dir + pid.replace("/", "_")
-        input_stream = io.open(path, "rb")
-        hex_digests, tmp_file_name = store.objects._mktempfile(input_stream)
-        input_stream.close()
-        assert hex_digests.get("md5") == pids[pid]["md5"]
-        assert hex_digests.get("sha1") == pids[pid]["sha1"]
-        assert hex_digests.get("sha256") == pids[pid]["sha256"]
-        assert hex_digests.get("sha384") == pids[pid]["sha384"]
-        assert hex_digests.get("sha512") == pids[pid]["sha512"]
-        assert os.path.isfile(tmp_file_name) is True
-
-
-def test_mktempfile_with_algorithm(pids, store):
-    test_dir = "tests/testdata/"
-    for pid in pids.keys():
-        path = test_dir + pid.replace("/", "_")
-        input_stream = io.open(path, "rb")
-        algo = "sha224"
-        hex_digests, tmp_file_name = store.objects._mktempfile(input_stream, algo)
-        input_stream.close()
-        assert hex_digests.get("sha224") == pids[pid]["sha224"]
-
-
-def test_mktempfile_with_unsupported_algorithm(pids, store):
-    test_dir = "tests/testdata/"
-    for pid in pids.keys():
-        path = test_dir + pid.replace("/", "_")
-        input_stream = io.open(path, "rb")
-        algo = "md2"
-        with pytest.raises(ValueError):
-            hex_digests, tmp_file_name = store.objects._mktempfile(input_stream, algo)
-        input_stream.close()
-
-
-def test_to_bytes(store):
-    string = "teststring"
-    string_bytes = store.objects._to_bytes(string)
-    assert isinstance(string_bytes, bytes)
-
-
-def test_get_store_path(store):
-    path_objects = store.objects._get_store_path()
-    path_objects_string = str(path_objects)
-    assert path_objects_string.endswith("/metacat/objects")
-    path_sysmeta = store.sysmeta._get_store_path()
-    path_sysmeta_string = str(path_sysmeta)
-    assert path_sysmeta_string.endswith("/metacat/sysmeta")
-
-
-def test_clean_algorithm(store):
-    algorithm_underscore = "sha_256"
-    algorithm_hyphen = "sha-256"
-    algorithm_other_hyphen = "sha3-256"
-    cleaned_algo_underscore = store.objects.clean_algorithm(algorithm_underscore)
-    cleaned_algo_hyphen = store.objects.clean_algorithm(algorithm_hyphen)
-    cleaned_algo_other_hyphen = store.objects.clean_algorithm(algorithm_other_hyphen)
-    assert cleaned_algo_underscore == "sha256"
-    assert cleaned_algo_hyphen == "sha256"
-    assert cleaned_algo_other_hyphen == "sha3_256"
